rename package pkg/eestream to uplink/eestream
