+++ pkg/audit/cursor.go
@@ -13,11 +13,11 @@ import (
 
 	"github.com/zeebo/errs"
 
-	"storj.io/storj/pkg/eestream"
 	"storj.io/storj/pkg/pb"
 	"storj.io/storj/pkg/storage/meta"
 	"storj.io/storj/pkg/storj"
 	"storj.io/storj/satellite/metainfo"
+	"storj.io/storj/uplink/eestream"
 )
 
 // Stripe keeps track of a stripe's index and its parent segment
