+++ satellite/metainfo/metainfo.go
@@ -21,7 +21,6 @@ import (
 	"storj.io/storj/pkg/accounting"
 	"storj.io/storj/pkg/auth"
 	"storj.io/storj/pkg/auth/signing"
-	"storj.io/storj/pkg/eestream"
 	"storj.io/storj/pkg/macaroon"
 	"storj.io/storj/pkg/overlay"
 	"storj.io/storj/pkg/pb"
@@ -31,6 +30,7 @@ import (
 	"storj.io/storj/satellite/console"
 	"storj.io/storj/satellite/orders"
 	"storj.io/storj/storage"
+	"storj.io/storj/uplink/eestream"
 )
 
 const (
