+++ uplink/eestream/common.go
