+++ uplink/eestream/decode.go
