+++ uplink/eestream/encode.go
