+++ uplink/eestream/rs.go
