+++ uplink/eestream/rs_test.go
